+++ mpisppy/utils/callbacks/termination/__init__.py
@@ -1,0 +1,5 @@
+# This software is distributed under the 3-clause BSD License.
+from .termination_callbacks import (
+    supports_termination_callback,
+    set_termination_callback,
+)
