+++ mpisppy/utils/callbacks/termination/tests/markshare2.py
